+++ Constants.lua
@@ -56,6 +56,8 @@ end
 Addon.ON_UPDATE_INTERVAL = 0.1 -- minimum number of seconds between each update of a frame for OnUpdate handlers
 Addon.CASTBAR_INTERRUPT_HOLD_TIME = 1
 
+Addon.NameplateParentFrame = WorldFrame
+Addon.NameplateFrameStrata = "BACKGROUND"
 Addon.UIScale = 1
 
 Addon.TotemInformation = {} -- basic totem information
@@ -562,8 +564,10 @@ Addon.DEFAULT_SETTINGS = {
     ShowFriendlyBlizzardNameplates = false,
     ShowEnemyBlizzardNameplates = false,
     Appearance = {
+      AnchorFrame = "WORLD_FRAME",
+      FrameStrata = "BACKGROUND",
       UseMasque = false,
-      UseBorderlessIcons = false,  
+      UseBorderlessIcons = false,
     },
     Automation = {
       FriendlyUnits = "NONE",